refactor: Move controller under adapter
