+++ backend_rust/src/adapters.rs
@@ -2,3 +2,4 @@ pub mod grpc;
 pub mod recipe_mapper;
 pub mod repository;
 pub mod step_mapper;
+pub mod controller;
