+++ backend_rust/src/adapters/controller.rs
@@ -1,0 +1,1 @@
+pub mod graphql;
