+++ backend_rust/src/adapters/controller/graphql.rs
@@ -1,0 +1,4 @@
+pub mod graphql_controller;
+pub mod mutation;
+pub mod object;
+pub mod query;
