+++ backend_rust/src/adapters/controller/graphql/graphql_controller.rs
@@ -1,7 +1,7 @@
 use async_graphql_axum::{GraphQLRequest, GraphQLResponse};
 use axum::extract::Extension;
 
-use crate::presentation::graphql::query::QuerySchema;
+use crate::adapters::controller::graphql::query::QuerySchema;
 
 pub async fn graphql_handler(
     schema: Extension<QuerySchema>,
