+++ backend_rust/src/adapters/step_mapper.rs
@@ -1,6 +1,6 @@
 use crate::application::mapper::api_mapper::ApiMapper;
 use crate::domain::Step;
-use crate::presentation::graphql::object::Step as StepObject;
+use crate::adapters::controller::graphql::object::Step as StepObject;
 pub struct StepMapper {}
 
 impl ApiMapper<Step, StepObject> for StepMapper {
