+++ backend_rust/src/bin/gen_schema.rs
@@ -1,7 +1,8 @@
 use async_graphql::{EmptySubscription, Schema};
-use cpp_backend::presentation::graphql::{mutation::Mutation, query::Query};
 use sea_orm::{ConnectOptions, Database};
 use std::env;
+use cpp_backend::adapters::controller::graphql::mutation::Mutation;
+use cpp_backend::adapters::controller::graphql::query::Query;
 
 #[tokio::main]
 async fn main() {
