+++ backend_rust/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod adapters;
 pub mod application;
 mod domain;
-pub mod presentation;
