--- backend_rust/src/presentation.rs
@@ -1,2 +1,0 @@
-pub mod controller;
-pub mod graphql;
