--- backend_rust/src/presentation/controller.rs
@@ -1,1 +1,0 @@
-pub mod graphql_controller;
