--- backend_rust/src/presentation/graphql.rs
@@ -1,3 +1,0 @@
-pub mod mutation;
-pub mod object;
-pub mod query;
